feat(remove): resources

+++ src/signati/index.ts
@@ -2,30 +2,24 @@ import * as crypto from 'crypto';
 import * as fs from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import {Concepts} from './tags/Concepts';
-import {Emisor} from './tags/Emisor';
-import {Impuestos} from './tags/Impuestos';
-import {Receptor} from './tags/Receptor';
-import {FileSystem} from './utils/FileSystem';
+import { Concepts } from './tags/Concepts';
+import { Emisor } from './tags/Emisor';
+import { Impuestos } from './tags/Impuestos';
+import { Receptor } from './tags/Receptor';
+import { FileSystem } from './utils/FileSystem';
 
-import {cer, key} from '@signati/openssl';
-import {saxon, Transform} from '@signati/saxon';
-import {js2xml} from 'xml-js';
-import {Relacionado} from './tags/Relacionado';
-import {ComlementType, XmlComplements} from './types/Tags/complements.interface';
-import {Comprobante, XmlComprobante, XmlComprobanteAttributes} from './types/Tags/comprobante.interface';
-import {XmlConcepto} from './types/Tags/concepts.interface';
-import {XmlCdfi, XmlVersion} from './types/Tags/xmlCdfi.interface';
-import {Structure} from './utils/structure';
-import {schema} from './utils/XmlHelp';
-import {TagComprobante} from "./types";
-
-interface Options {
-    debug: boolean;
-    compact?: boolean;
-    customTags?: any;
-}
-
+import { cer, key } from '@signati/openssl';
+import { saxon, Transform } from '@signati/saxon';
+import { js2xml } from 'xml-js';
+import { Relacionado } from './tags/Relacionado';
+import { ComlementType, XmlComplements } from './types/Tags/complements.interface';
+import { Comprobante, XmlComprobante, XmlComprobanteAttributes, XmlnsLinks } from './types/Tags/comprobante.interface';
+import { XmlConcepto } from './types/Tags/concepts.interface';
+import { XmlCdfi, XmlVersion } from './types/Tags/xmlCdfi.interface';
+import { Structure } from './utils/structure';
+import { schema } from './utils/XmlHelp';
+import { TagComprobante } from "./types";
+import { Options } from './types/types';
 export class CFDI {
     private xml: XmlCdfi = {} as XmlCdfi;
     private debug: boolean = false;
@@ -33,32 +27,57 @@ export class CFDI {
     private version: string = '3.3';
     private tags: Structure;
     private tc: TagComprobante = 'cfdi:Comprobante';
-
-    constructor(attribute: Comprobante, options: Options = {
-        debug: false,
-        compact: false,
-        customTags: {}
-    }) {
-        this.tags = new Structure(options.customTags)
+    private XMLSchema = 'http://www.w3.org/2001/XMLSchema-instance';
+    private cfd = 'http://www.sat.gob.mx/cfd/4';
+    private locations = [
+        'http://www.sat.gob.mx/cfd/4',
+        'http://www.sat.gob.mx/sitio_internet/cfd/4/cfdv40.xsd',
+    ];
+    private xslt: string | null = null;
+    constructor(attribute: Comprobante, options: Options) {
+        const { debug = false, compact = false, xslt, customTags = {} } = options
+        xslt && (this.xslt = xslt);
+        this.tags = new Structure(customTags)
         this.tc = this.tags.tagXml<TagComprobante>('cfdi:Comprobante')
-        this.debug = options.debug
-
+        this.debug = debug
         this.restartCfdi();
-        this.addXmlns('xsi', 'http://www.w3.org/2001/XMLSchema-instance')
-        this.addXmlns('cfdi', 'http://www.sat.gob.mx/cfd/3')
-
-        this.addSchemaLocation([
-            'http://www.sat.gob.mx/cfd/3',
-            'http://www.sat.gob.mx/sitio_internet/cfd/3/cfdv33.xsd',
-        ]);
+        this.xmlns(attribute.xmlns || { xsi: this.XMLSchema, cfdi: this.cfd })
+        this.addSchemaLocation(attribute.schemaLocation || this.locations);
+        attribute.xmlns && delete attribute.xmlns
+        attribute.schemaLocation && delete attribute.schemaLocation
         this.xml[this.tc]._attributes = {
             ...this.xml[this.tc]._attributes,
-            ...{Version: this.version},
+            ...{ Version: this.version },
             ...attribute
         };
     }
 
-    public async setAttributesXml(attribute: XmlVersion = {version: '1.0', encoding: 'utf-8'}) {
+    private xmlns(xmlns: XmlnsLinks) {
+        if (!xmlns.xsi) {
+            this.addXmlns('xsi', this.XMLSchema)
+        }
+        if (!xmlns.cfdi) {
+            this.addXmlns('cfdi', this.cfd)
+        }
+
+        for (const xmln in xmlns) {
+            this.addXmlns(xmln, xmlns[xmln])
+        }
+    }
+    private addXmlns(xmlnsKey: string, xmlns: string) {
+        this.xml[this.tc]._attributes['xmlns:' + xmlnsKey] = xmlns;
+    }
+
+    private addSchemaLocation(locations: string[]) {
+
+        if (!this.xml[this.tc]._attributes['xsi:schemaLocation']) {
+            this.xml[this.tc]._attributes['xsi:schemaLocation'] = '';
+        }
+        const schemaLocation = schema(locations);
+        this.xml[this.tc]._attributes['xsi:schemaLocation'] += ' ' + schemaLocation;
+    }
+
+    public async setAttributesXml(attribute: XmlVersion = { version: '1.0', encoding: 'utf-8' }) {
         this.xml._declaration._attributes = attribute;
     }
 
@@ -66,14 +85,14 @@ export class CFDI {
     public setAttributesComprobantes(attribute: Comprobante) {
         this.xml[this.tc]._attributes = {
             ...this.xml[this.tc]._attributes,
-            ...{Version: this.version},
+            ...{ Version: this.version },
             ...attribute
         };
 
     }
 
     public async relacionados(relationCfdi: Relacionado) {
-        this.xml[this.tc] = Object.assign({['cfdi:CfdiRelacionados']: relationCfdi.getRelation()}, this.xml[this.tc]);
+        this.xml[this.tc] = Object.assign({ ['cfdi:CfdiRelacionados']: relationCfdi.getRelation() }, this.xml[this.tc]);
     }
 
     public async emisor(emisor: Emisor) {
@@ -144,22 +163,10 @@ export class CFDI {
      * @param {String} password
      */
     public async sellar(keyfile: string, password: string) {
-        try {
-            // console.log('sellar')
-            const cadena = await this.getCadenaOriginal();
-            // console.log('caenda', cadena)
-            const sello = await this.getSello(cadena, keyfile, password);
-            // console.log('sello', sello)
-            this.xml[this.tc]._attributes.Sello = sello;
-        } catch (e) {
-            if (this.debug) {
-                console.log({
-                    method: 'getCadenaOriginal',
-                    error: e
-                })
-            }
-            return e.messageundefined
-        }
+        const cadena = await this.getCadenaOriginal();
+        const sello = await this.getSello(cadena, keyfile, password);
+        this.xml[this.tc]._attributes.Sello = sello;
+
     }
 
     public async getJsonCdfi(): Promise<XmlCdfi> {
@@ -167,7 +174,7 @@ export class CFDI {
             try {
                 resolve(this.xml);
             } catch (e) {
-                reject({message: e});
+                reject({ message: e });
             }
         });
 
@@ -176,12 +183,12 @@ export class CFDI {
     public async getXmlCdfi(): Promise<string> {
         return await new Promise(async (resolve, reject) => {
             try {
-                const options = {compact: true, ignoreComment: true, spaces: 4};
-                const cfdi = await js2xml({...this.xml}, options);
+                const options = { compact: true, ignoreComment: true, spaces: 4 };
+                const cfdi = await js2xml({ ...this.xml }, options);
                 this.restartCfdi();
                 resolve(cfdi);
             } catch (e) {
-                reject({message: e});
+                reject({ message: e });
             }
         });
 
@@ -223,38 +230,24 @@ export class CFDI {
 
     }
 
-    private async addXmlns(xmlnsKey: string, xmlns: string) {
-        this.xml[this.tc]._attributes['xmlns:' + xmlnsKey] = xmlns;
-    }
-
-    private async addSchemaLocation(locations: string[]) {
-
-        if (!this.xml[this.tc]._attributes['xsi:schemaLocation']) {
-            this.xml[this.tc]._attributes['xsi:schemaLocation'] = '';
-        }
-        const schemaLocation = schema(locations);
-        this.xml[this.tc]._attributes['xsi:schemaLocation'] += ' ' + schemaLocation;
-    }
-
     private async getCadenaOriginal(): Promise<string> {
+        if (!this.xslt) {
+            throw new Error('¡Ups! Direcction Not Found Extensible Stylesheet Language Transformation')
+        }
         return new Promise(async (resolve, reject) => {
             try {
                 const fullPath = path.join(os.tmpdir(), `${FileSystem.generateNameTemp()}.xml`);
-                const options = {compact: true, ignoreComment: true, spaces: 4};
+                const options = { compact: true, ignoreComment: true, spaces: 4 };
                 const result = js2xml(this.xml, options);
                 fs.writeFileSync(fullPath, result, 'utf8');
-                const stylesheetDir = path.join(path.resolve(__dirname, '../signati'), 'resources/xslt33/', 'cadenaoriginal_3_3.xslt');
-                if (this.debug) {
-                    console.log(stylesheetDir);
-                }
                 const transform = new Transform();
-                const cadena = transform.s(fullPath).xsl(stylesheetDir).warnings('silent').run();
-                // const cadena = saxon(stylesheetDir, fullPath);
+                const cadena = transform.s(fullPath).xsl(String(this.xslt)).warnings('silent').run();
 
                 if (this.debug) {
                     /*
                     * ||3.3|E|ACACUN-27|2014-07-08T12:16:50|Pago en una sola exhibición|20001000000300022815|16148.04|645.92|MXN|17207.35|I|En efectivo|México|01|asdasd-3234-asdasd-2332-asdas|asdasd-3234-asdasd-2332-asdas|TCM970625MB1|FACTURACION MODERNA SA DE CV|601|XAXX010101000|PUBLICO EN GENERAL|G01|001|1212|2|pieza|Pieza|audifonos|1000|2000|00.0|369.83|002|Tasa|0.16|59.17|369.8aaaa3|002|Tasa|0.16|59.17|369.83|002|Tasa|0.16|59.17|002|59.17|1000|002|Tasa|0.16|59.17||
                     * */
+                    console.log(this.xslt);
                     console.log('cadena original =>', cadena)
                 }
                 fs.unlinkSync(fullPath);
@@ -267,7 +260,7 @@ export class CFDI {
                         error: e
                     })
                 }
-                reject({message: e});
+                reject({ method: 'getCadenaOriginal', message: e });
             }
         });
 
@@ -289,7 +282,7 @@ export class CFDI {
                         error: e
                     })
                 }
-                reject({message: e});
+                reject({ message: e });
             }
         });
     }
